refact(service): Preparing user service for jwt and another stuffs

+++ src/main/java/com/combos/TaskManager/service/UserService.java
@@ -51,31 +51,33 @@ public class UserService {
         return UserMapper.toDTO(user);
     }
 
-    public UserResponseDTO updateUser(Long id, UserUpdateDTO dto) {
+    public UserResponseDTO updateUser(Long requesterUserId, Long id, UserUpdateDTO dto) {
         User user = findUserById(id);
+
+        if (!user.getId().equals(requesterUserId)) {
+            throw new RuntimeException("You can only update your own account");
+        }
 
         if (dto.name() != null) {
             user.setName(dto.name());
         }
 
-        if (dto.nickname() != null) {
-            if (!user.getNickname().equals(dto.nickname()) &&
-                repository.existsByNickname(dto.nickname())) {
+        if (dto.nickname() != null && !dto.nickname().equals(user.getNickname())) {
+            if (repository.existsByNickname(dto.nickname())) {
                 throw new RuntimeException("Nickname already in use");
             }
             user.setNickname(dto.nickname());
         }
 
-        if (dto.email() != null) {
-            if (!user.getEmail().equals(dto.email()) &&
-                repository.existsByEmail(dto.email())) {
+        if (dto.email() != null && !dto.email().equals(user.getEmail())) {
+            if (repository.existsByEmail(dto.email())) {
                 throw new RuntimeException("Email already in use");
             }
             user.setEmail(dto.email());
         }
 
         if (dto.password() != null) {
-            user.setPassword(dto.password());
+            user.setPassword(dto.password()); // TODO: Password encoder here
         }
 
         user = repository.save(user);
@@ -83,8 +85,11 @@ public class UserService {
         return UserMapper.toDTO(user);
     }
 
-    public void deleteById(Long id) {
+    public void deleteById(Long requesterUserId, Long id) {
         User user = findUserById(id);
+        if (!user.getId().equals(requesterUserId)) {
+            throw new RuntimeException("You can only delete your own account");
+        }
         repository.delete(user);
     }
 }
